Added Strategy::Base#successful? to check whether the strategy is a winning one or not.
Previously, if a strategy would have a user set, Warden wouldn't check
at the result to see if the strategy is a :success. It merely look at
the strategy#user parameter and expect it to be succesful if a user is
set.

This can lead to unexpected result if a strategy would use the #user
attribute to use memoization while it checks for if the authentication
is successful.