Make it able to throw in Warden::Manager.on_request
In `Warden::Test::Helpers#login_as`, it's using
`Warden.on_next_request`, which is using
`Warden::Manager.on_request`, and it would call
`proxy.set_user` which would possibly throw.

This means before this patch, `Proxy.new` could
throw in the tests. We're also expecting `env['warden']`
to be set in `Warden::Manager`, therefore we need to
decouple initializing the object and calling the `on_request`
callbacks, and move `env['warden'].on_request` inside the
`catch` block to handle the logging in process properly
in the tests.

I am not sure if people are relying on this behaviour
other than tests, but I assume it's not bad to be able to
`throw` in `Warden::Manager.on_request` as a way to
stop the chain early.

Without this patch, in the tests, `:warden` would be thrown
whenever it failed to login. People might be depending on
this behaviour, but throwing symbols like this could break
people's cleanup code.

Here's an example: https://github.com/rails/rails/pull/28373

Rails was using `rescue Exception` before to cleanup stuffs,
but it would not be rescuing any symbols. There were other
codes doing things like this, causing some weird transient
test failures like this for GitLab:
https://gitlab.com/gitlab-org/gitlab-ce/issues/40093

In short, we shouldn't let any `throw` bubble up. The other
way to fix this would be not using `Warden::Manager.on_request`
to login, but that might need more changes and could break
even more codes. Also, I think it's not bad to be able to
`throw :warden` in `Warden::Manager.on_request`.